+++ pkg/backend/display/events.go
@@ -56,13 +56,14 @@ func ConvertEngineEvent(e engine.Event) (apitype.EngineEvent, error) {
 			return apiEvent, eventTypePayloadMismatch
 		}
 		apiEvent.PolicyEvent = &apitype.PolicyEvent{
-			ResourceURN:       string(p.ResourceURN),
-			Message:           p.Message,
-			Color:             string(p.Color),
-			PolicyName:        p.PolicyName,
-			PolicyPackName:    p.PolicyPackName,
-			PolicyPackVersion: p.PolicyPackVersion,
-			EnforcementLevel:  string(p.EnforcementLevel),
+			ResourceURN:          string(p.ResourceURN),
+			Message:              p.Message,
+			Color:                string(p.Color),
+			PolicyName:           p.PolicyName,
+			PolicyPackName:       p.PolicyPackName,
+			PolicyPackVersion:    p.PolicyPackVersion,
+			PolicyPackVersionTag: p.PolicyPackVersion,
+			EnforcementLevel:     string(p.EnforcementLevel),
 		}
 
 	case engine.PreludeEvent: