+++ tests/integration/custom_timeouts/custom_timeouts_test.go
@@ -9,25 +9,23 @@ import (
 
 func TestCustomTimeouts(t *testing.T) {
 	opts := &integration.ProgramTestOptions{
-		Dir: filepath.Join(".", "python"),
+		Dir: filepath.Join(".", "python", "success"),
 		Dependencies: []string{
 			filepath.Join("..", "..", "..", "sdk", "python", "env", "src"),
 		},
-		Quick: true,
-		EditDirs: []integration.EditDir{
-			{
-				Dir:      "step1",
-				Additive: true,
-			},
-			{
-				Dir:      "step2",
-				Additive: true,
-			},
-			{
-				Dir:      "step3",
-				Additive: true,
-			},
+		Quick:      true,
+		NoParallel: true,
+	}
+	integration.ProgramTest(t, opts)
+
+	opts = &integration.ProgramTestOptions{
+		Dir: filepath.Join(".", "python", "failure"),
+		Dependencies: []string{
+			filepath.Join("..", "..", "..", "sdk", "python", "env", "src"),
 		},
+		Quick:         true,
+		NoParallel:    true,
+		ExpectFailure: true,
 	}
 	integration.ProgramTest(t, opts)
 }